🔥 Remove update system patch
This is causing issues upstream (pulse-browser/browser#173). This wouldn't be an issue if mach respected overriding `@MOZ_APPUPDATE_HOST@`, but that does not appear to be the case. Instead, this patch should be handled by upstream users.

If anyone needs help migrating, open an issue or ping me in discord